requirements.txt
numpy>=1.21.0
gymnasium>=0.29.0
pillow>=9.0.0
opencv-python>=4.5.0
pyyaml>=6.0

# Testing (optional)
pytest>=7.0.0

# Type checking (optional)
# mypy>=1.0.0
